refactor: ページ構成を1ページに統合
- src/app/check-state/page.tsx を削除
- src/app/setting/page.tsx を削除
- calcFocusData.tsx 内の不要なダミー変数を削除
